+++ src/main/java/org/fundacionjala/converter/controller/request/RequestExtractMetadataParam.java
@@ -7,33 +7,14 @@
  * license agreement you entered into with Fundacion Jala
  */
 package org.fundacionjala.converter.controller.request;
-import org.springframework.web.multipart.MultipartFile;
 
 /**
  * @author Angela Martinez
  * @version 0.1
  */
-public class RequestExtractMetadataParam {
-    private MultipartFile file;
-    private String exportFormat;
+public class RequestExtractMetadataParam extends RequestParameter {
     private String detail;
     private String checksumMD5;
-
-    /**
-     * Sets file
-     * @param file
-     */
-    public void setFile(final MultipartFile file) {
-        this.file = file;
-    }
-
-    /**
-     * Returns path of file
-     * @return path
-     */
-    public MultipartFile getFile() {
-            return file;
-        }
 
     /**
      * Returns checksumMD5
@@ -52,27 +33,11 @@ public class RequestExtractMetadataParam {
     }
 
     /**
-     * Sets format to exports metadata
-     * @param exportFormat
-     */
-    public void setExportFormat(final String exportFormat) {
-        this.exportFormat = exportFormat;
-    }
-
-    /**
      * Set amount of detail of metadata
      * @param detail
      */
     public void setDetail(final String detail) {
         this.detail = detail;
-    }
-
-    /**
-     * Returns format to export metadata
-     * @return format
-     */
-    public String getExportFormat() {
-        return exportFormat;
     }
 
     /**
